tweak: Dynamically adjustable clamping of smart streaming statistics
The added tweaks make it possible do dynamically clamp the input
metrics to avoid bogus metrics. This is to protect against transient
values that were observed experimentally for example on FPS due to
effects such as TCP packet reordering by the client-side kernel
(causing the incoming FPS to become noisy and jump back and forth,
e.g. you'd observe 0 then 100).

This is a temporary solution. A better approach would have to be a
slow-moving min/max window with some averaging. To be done later.