Display channel implementation of the "smart-streaming" core function
The dcc_handle_stream_metrics function receives new metrics from client,
then calls dcc_process_metrics to process the received metrics and
generate adjustments to the encoder.

In the long term, we want to unify the stream and DCC versions of that
function with a single "core" function, and using the dcc_process_metrics
function as a mere wrapper around that core function.