Restore original (following) algorithm in new infrastructure
This is an alternative smart-streaming algorithm that is conceptually
simpler, and is what was originally described. It basically consists
in attempting to follow whatever is measured client-side.
There is an improvement over the original idea which is go give
different weights to the measurements at various stages of the decoding
pipeline.

A selection of smart-streaming algorithm can be made using smstr_algo:

- smstr_algo = 0 (default): No smart streaming
- smstr_algo = 1: Following algorithm (original method)
- smstr_algo = 2: Thinking algorithm (new method)